Module 2/Chapter11/django-myproject-11/requirements.txt
Django==1.8.5
django-crispy-forms
requests

djangorestframework
markdown
django-filter

selenium
mock

cookiecutter
